+++ components/SearchDetails.js
@@ -8,11 +8,11 @@ import styles from '../styles/SearchDetails.module.css'
 const SearchDetails = ({ list }) => {
   const searchList = list && arrayHandler(list)
   return (
-        <>
-            <div className={styles.contextWrapper}>
-                {searchList.map((el, index) => <span role='text' key={index} className={styles.context}>{el}</span>)}
-            </div>
-        </>
+    <>
+        <div className={styles.contextWrapper}>
+            {searchList.map((el, index) => <span role='text' key={index} className={styles.context}>{el}</span>)}
+        </div>
+    </>
   )
 }
